tests: reorganize the test classes for a nicer layout
